refactor: top level scripts
