+++ README.md
@@ -30,7 +30,7 @@ The scripts have been implemented to be run inside the Docker container only. Th
 docker build --tag=topo-imagery .
 ```
 
-- Example: running `standardising_validate.py` script
+- Example: running the `standardise-validate` command
 
 This script standardises TIFF files to [COGs](https://www.cogeo.org/) with a creation of a [STAC](https://stacspec.org/) Item file per TIFF containing the metadata.
 The input TIFF file paths have to be passed through a `json` file in the following format:
@@ -48,12 +48,12 @@ where `output` is the desired output [tile name](https://github.com/linz/topo-im
 
 Some test data are available in `/scripts/tests/data/` along with the expected output.
 
-Run `docker run topo-imagery python standardise_validate.py --help` to get the list of the expected arguments.
+Run `docker run topo-imagery standardise-validate --help` to get the list of the expected arguments.
 
 - Example of local execution. This example uses the test data available on this repo and create the output will be created in a `~/tmp/` on the local machine (volume share with `Docker`):
 
 ```bash
-docker run -v ${HOME}/tmp/:/tmp/:rw topo-imagery python standardise_validate.py --preset webp --from-file ./tests/data/aerial.json --collection-id 123 --start-datetime 2023-01-01 --end-datetime 2023-01-01 --target /tmp/ --source-epsg 2193 --target-epsg 2193 --gsd 10 --create-footprints=true
+docker run -v ${HOME}/tmp/:/tmp/:rw topo-imagery standardise-validate --preset webp --from-file ./tests/data/aerial.json --collection-id 123 --start-datetime 2023-01-01 --end-datetime 2023-01-01 --target /tmp/ --source-epsg 2193 --target-epsg 2193 --gsd 10 --create-footprints=true
 ```
 
 To use an AWS test dataset (input located in an AWS S3 bucket), log into the AWS account and add the following arguments to the `docker run` command:
