+++ packages/topo-imagery-gdal/src/topo_imagery_gdal/thumbnails.py
