+++ packages/topo-imagery-stac/src/topo_imagery_stac/generate_hillshade.py
