--- scripts/conftest.py
@@ -1,1 +1,0 @@
-from topo_imagery_stac.testing.helpers import fake_collection_context  # pylint: disable=unused-import
